change counting number of cpus for container

+++ operating_systems/abstract_operating_system.py
@@ -141,7 +141,7 @@ class AbstractOSFuncs:
         pass
 
     @abstractmethod
-    def get_container_number_of_cores(self) -> int:
+    def get_container_number_of_cores(self) -> float:
         pass
 
     @abstractmethod
