+++ operating_systems/os_windows.py
@@ -208,7 +208,7 @@ class WindowsOS(AbstractOSFuncs):
     def get_container_total_cpu_usage(self) -> float:
         raise NotImplementedError("Not implemented total cpu for windows container")
 
-    def get_container_number_of_cores(self) -> int:
+    def get_container_number_of_cores(self) -> float:
         raise NotImplementedError("Not implemented number of cores for windows container")
 
     def get_container_total_memory_usage(self) -> tuple[float, float]:
